unify has active calls logic by adding helper function
