+++ packages/ringcentral-widgets/containers/DialerPage/index.js
@@ -2,15 +2,16 @@ import { connect } from 'react-redux';
 import formatNumber from 'ringcentral-integration/lib/formatNumber';
 import callingModes from 'ringcentral-integration/modules/CallingSettings/callingModes';
 
+import { withPhone } from '../../lib/phoneContext';
+import hasActiveCalls from '../../lib/hasActiveCalls';
 import DialerPanel from '../../components/DialerPanel';
 import styles from './styles.scss';
-import { withPhone } from '../../lib/phoneContext';
 
 function mapToProps(_, {
+  phone,
   phone: {
     call,
     dialerUI,
-    callMonitor,
     callingSettings,
     contactSearch,
     connectivityMonitor,
@@ -18,7 +19,6 @@ function mapToProps(_, {
     rateLimiter,
     webphone,
     audioSettings,
-    conferenceCall,
   },
   dialButtonMuted = false,
 }) {
@@ -26,13 +26,7 @@ function mapToProps(_, {
   const waitingWebphoneConnected = (isWebphoneMode && webphone && webphone.connecting);
   const webphoneDisconnected = (isWebphoneMode && webphone && !webphone.connected);
   const audioNotEnabled = isWebphoneMode && audioSettings && !audioSettings.userMedia;
-  const conferenceCallEquipped = !!conferenceCall;
-  const withTab = !!(
-    conferenceCallEquipped
-    && isWebphoneMode
-    && callMonitor.calls.length
-    && webphone.sessions.length
-  );
+  const withinTab = hasActiveCalls(phone);
 
   return {
     currentLocale: locale.currentLocale,
@@ -61,7 +55,7 @@ function mapToProps(_, {
     dialButtonVolume: audioSettings ? audioSettings.dialButtonVolume : 1,
     // If audioSettings is used, then use values from audioSettings module
     dialButtonMuted: audioSettings ? audioSettings.dialButtonMuted : dialButtonMuted,
-    callBtnClassName: withTab ? null : styles.callBtn,
+    callBtnClassName: withinTab ? null : styles.callBtn,
   };
 }
 function mapToFunctions(_, {
